+++ src/minisphere/main.c
@@ -241,8 +241,6 @@ main(int argc, char* argv[])
 	initialize_debugger(want_debug, false);
 #endif
 
-	console_log(0, "engine started!");
-
 	// display loading message, scripts may take a bit to compile
 	al_clear_to_color(al_map_rgba(0, 0, 0, 255));
 	screen_draw_status(g_screen, "starting up...");