+++ src/plugin/PluginMain.cs
@@ -128,7 +128,7 @@ namespace minisphere.Gdk
 
         public int Verbosity
         {
-            get { return Math.Min(Math.Max(Conf.GetInteger("verbosity", 1), 0), 4); }
+            get { return Math.Min(Math.Max(Conf.GetInteger("verbosity", 2), 0), 4); }
             set { Conf.SetValue("verbosity", Math.Min(Math.Max(value, 0), 4)); }
         }
     }